fix: case-sensitive import path issues by renaming files
