+++ src/lib/tools/Tool.interface.ts
